feat: Add coins per kill
- Also changes purchasing so you can pick what you want to build instead
  of only allowing turret

+++ game/Assets/Scripts/Bugs/Bug.cs
@@ -6,7 +6,7 @@ using UnityEngine.Video;
 
 public class Bug : MonoBehaviour
 {
-    public int HitPoints = 2;
+    public float HitPoints = 2;
     
     [SerializeField] private float _movementSpeed = 2;
